+++ dojo_manager.py
@@ -20,122 +20,8 @@ arguments:
 
 import docopt
 import random
-
-
-class Room(object):
-
-    '''
-    base class Room to instantiate and hold attributes for the facilities at the dojo
-    '''
-    # room_name = ''
-
-    def __init__(self, room_name='', room_type=''):
-        self.room_name = room_name
-        self.room_type = room_type
-        all_rooms = []
-
-
-    def create_room(self, room_name, room_type):
-        self.room_name = room_name
-        self.room_type = room_type
-        print('Making Room Now')
-        self.all_rooms.append(self)
-
-        return self
-
-
-class Person(object):
-    '''
-    Base class to manage the people objects in the Dojo
-    '''
-
-    def __init__(self, name='', job_group='', accommodation=False):
-
-        self.person_name = name
-        self.accommodation = accommodation
-        self.job_group = job_group
-
-    def add_person(self, name, job_group, accommodation):
-        self.person_name = name
-        self.accommodation = accommodation
-        return self
-
-
-class Fellow(Person):
-    '''
-    Subclass Fellow to model Andela Fellows
-    '''
-
-    def __init__(self, fellow_name='', accomodation=False):
-        # Person().__init__(fellow_name, accomodation)
-        self.fellow_name = fellow_name
-        self.accommodation = accomodation
-
-    def add_person(self, fellow_name, accomodation):
-        self.fellow_name = fellow_name
-        self.accommodation = accomodation
-        print('Fellow {0} {1} has been successfully added.'.format(
-            self.fellow_name[0], self.fellow_name[1]))
-        if not accomodation:
-            print('{0} does not wish to be accomodated'.format(fellow_name[0]))
-
-        return self
-
-
-class Staff(Person):
-    '''
-    subclass of Person to model staff members
-    '''
-
-    def __init__(self, staff_name=''):
-        self.person_name = staff_name
-
-    def add_person(self, staff_name=''):
-        self.person_name = staff_name
-        print('Staff {0} {1} has been successfully added.'.format(
-            staff_name[0], staff_name[1]))
-        print(self.person_name)
-        return self
-
-
-class Office(Room):
-
-    '''
-    Subclass of Room to model offices
-    with attributes room_name and room_type
-    '''
-    def __init__(self, room_name='', room_type=''):
-        self.room_name = room_name
-        self.room_type = room_type
-
-    def create_room(self, room_name, room_type):
-        self.room_name = room_name
-        self.room_type = room_type
-        return self
-
-
-class LivingSpace(Room):
-    '''
-    Subclass of Room to model living spaces with attributes room_name and
-    room_type
-    '''
-
-    def __init__(self, room_name='', room_type=''):
-        self.room_name = room_name
-        self.room_type = room_type
-        self.occupants = []
-
-
-    def create_room(self, room_name, room_type):
-
-        self.room_name = room_name
-        self.room_type = room_type
-
-        return self
-    #
-    # def add_occupant(self, person):
-    #     self.occupants.append(person)
-
+from personel.person import Person, Staff
+from rooms.room import Office, LivingSpace
 
 class Dojo(object):
     '''
@@ -208,10 +94,10 @@ class Dojo(object):
 
 
 #Print for testing purposes.
-        print(len(self.livingspaces))
-        print(len(self.offices))
-        print(len(self.fellows))
-        print(len(self.staff))
+        # print(len(self.livingspaces))
+        # print(len(self.offices))
+        # print(len(self.fellows))
+        # print(len(self.staff))
 
 
     def add_office(self, room_name):
